modified:   .config/glaze-wm/home.yaml 	modified:   .config/komorebi/komorebi.ps1 	modified:   .config/whkdrc 	modified:   init.ps1 	modified:   powershell/Scripts/powershell.check.ps1 	modified:   powershell/Scripts/powershell.mainload.ps1 	modified:   powershell/profile.ps1 	new file:   userconfig.json 	deleted:    userconfig.psd1

+++ init.ps1
@@ -1,35 +1,48 @@
-# general variable
-    $ErrorActionPreference = 'Ignore'
+# default error handling
+$ErrorActionPreference = 'Stop'
+# get userconfig
+$config = Get-Content $PSScriptRoot\userconfig.json | ConvertFrom-Json
+# check fo installed software
+[hashtable]$checked = & $PSScriptRoot\powershell\Scripts\powershell.check.ps1
+# path to the startup folder
+[string]$startupPath = "$env:APPDATA\Microsoft\Windows\Start Menu\Programs\Startup"
+# init arraylists
+$shortcuts = New-Object System.Collections.ArrayList
+$junctions = New-Object System.Collections.ArrayList
 
-# import userconfig
-    $config = Import-PowerShellDataFile $PSScriptRoot\userconfig.psd1 -ErrorAction Stop
+function New-Shortcut {
+    param(
+        [string]
+        # Specifies the file name
+        $name,
+        # Spcifies the target of the shortcut
+        $target,
+        # add arguments if needed
+        $arguments,
+        # path to where the shortcut should be generated
+        $path
+    )
 
-# check for installed software
-    $checked = & $PSScriptRoot\powershell\Scripts\powershell.check.ps1
-
-# path to windows startup folder
-    $startupPath = "$env:APPDATA\Microsoft\Windows\Start Menu\Programs\Startup"
-
-# init $shortcuts as arraylist
-    $shortcuts = [System.Collections.ArrayList]::new()
-
-# init $junctions as arraylist
-    $junctions = [System.Collections.ArrayList]::new()
-
-# scriptblock for creating a shortcut
-    $createLink = {
-        param( $name,$sourceExe,$argumentsExe,$destination )
+    process {
         $WshShell = New-Object -comObject WScript.Shell
-        $shortcut = $WshShell.CreateShortcut("$destination\$name.lnk")
-        $shortcut.TargetPath = $sourceExe
-        $shortcut.Arguments = $argumentsExe
+        $shortcut = $WshShell.CreateShortcut("$path\$name.lnk")
+        $shortcut.TargetPath = $target
+        $shortcut.Arguments = $arguments
         $shortcut.Save()
     }
+}
 
-# Script block to copy old files to tmp and put them back into the target of junction
-    $createJunction = {
-        param( $path,$target )
-        $guid = New-Guid
+function New-Junction {
+    param(
+        [string]
+        # original path of the directory
+        $path,
+        # target path of the junction
+        $target
+    )
+    
+    process {
+        $guid = [guid]::newguid()
         $tmp = "$env:TEMP\$guid"
         if (!(Test-Path -Path $path)) {
             $null = New-Item -ItemType Junction -Path $path -Target $target
@@ -47,12 +60,39 @@
             Remove-Item -Path $tmp -Recurse -Force
         }
     }
+}
 
-# error out if missing software 
+function Out-Error {
+    param(
+        [string]
+        # error message
+        $msg,
+        # error escalation
+        $escnum
+    )
+
+    switch($escnum) {
+        0 {
+            Write-Warning $msg
+        }
+        1{
+            Write-Error $msg
+        }
+        2 {
+            Write-Error $msg
+            Exit
+        }
+        default {
+            Write-Debug 'Something wrong with error handling'
+        }
+    }
+}
 
 # take actions on stuff in the userconfig file
+
 # windowmanager
-    if ($config.windowmanager -eq 'glazewm') {
+# glazewm
+    if ($config.windowmanager -eq 'glazewm' -and $checked.GlazeWM) {
         # check for komorebi shortcut and delete it
             if (Test-Path "$startupPath\komorebi.lnk") {
                 Remove-Item -Path "$startupPath\komorebi.lnk"
@@ -60,12 +100,15 @@
         # add glazewm to $shortcuts
             [void]$shortcuts.Add(@{
                 name = 'glazewm'
-                sourceExe = "powershell.exe"
-                argumentsExe = "-WindowStyle hidden -Command glazewm --config=$env:USERPROFILE\.config\glaze-wm\$($config.workspace).yaml"
-                destination = $startupPath
+                target = "powershell.exe"
+                arguments = "-WindowStyle hidden -Command glazewm --config=$env:USERPROFILE\.config\glaze-wm\$($config.workspace).yaml"
+                path = $startupPath
             })
-    } 
-    else {
+        # copy .config files
+            Copy-Item -Path $PSScriptRoot\.config\glaze-wm -Destination $env:USERPROFILE\.config -Recurse -Force
+    }
+# komorebi
+    elseif ($config.windowmanager -eq 'komorebi' -and $checked.Komorebi) {
         # check for glazewm shortcut and delete it
             if (Test-Path "$startupPath\glazewm.lnk") {
                 Remove-Item -Path "$startupPath\glazewm.lnk"
@@ -73,59 +116,116 @@
         # add komorebi to $shortcuts
             [void]$shortcuts.Add(@{
                 name = 'komorebi'
-                sourceExe = "powershell.exe"
-                argumentsExe = '-WindowStyle hidden -Command komorebic start --await-configuration'
-                destination = $startupPath
+                target = "powershell.exe"
+                arguments = '-WindowStyle hidden -Command komorebic start --await-configuration'
+                path = $startupPath
             })
+        # copy .config files
+            Copy-Item -Path $PSScriptRoot\.config\komorebi -Destination $env:USERPROFILE\.config -Recurse -Force
     }
-
-# taskbar setting
-    if ($config.taskbar) {
-        # add nircmd to $shortcuts with 255
-            [void]$shortcuts.Add(@{
-                name = 'nircmd'
-                sourceExe = (Get-Command nircmd).Path
-                argumentsExe = 'win trans class Shell_TrayWnd 255'
-                destination = $startupPath
-            })
-    } 
+# error handling
     else {
-        # add nircmd to $shortcuts with 256
-            [void]$shortcuts.Add(@{
-                name = 'nircmd'
-                sourceExe = (Get-Command nircmd).Path
-                argumentsExe = 'win trans class Shell_TrayWnd 256'
-                destination = $startupPath
-            })
+        if (!$checked.Komorebi -and $config.windowmanager -eq 'komorebi') {
+            Out-Error -msg 'Missing Komorebi' -escnum 1 
+        }
+        if (!$checked.GlazeWM -and $config.windowmanager -eq 'glazewm') {
+            Out-Error -msg 'Missing Glaze-WM' -escnum 1
+        }
+        if ($config.windowmanager -ne ('komorebi' -or 'glazewm')) {
+            Out-Error -msg "Incorrect name of Windowmanager: $($config.windowmanager)" -escnum 0
+        }
     }
 
-# add WindowsTerminal localstate folder to $junctions
+# taskbar
+# Check for bool type
+    if ($config.taskbar.GetType().name -ne 'Boolean') {
+        Out-Error -msg "Incorrect data type of taskbar setting: $($config.taskbar)" -escnum 2
+    }
+# show taskbar
+    elseif ($config.taskbar -and $checked.Nircmd) {
+        [void]$shortcuts.Add(@{
+            name = 'nircmd'
+            target = (Get-Command nircmd).Path
+            arguments = 'win trans class Shell_TrayWnd 255'
+            path = $startupPath
+        })
+    }
+# remove taskbar
+    elseif (!$config.taskbar -and $checked.Nircmd) {
+        [void]$shortcuts.Add(@{
+            name = 'nircmd'
+            target = (Get-Command nircmd).Path
+            arguments = 'win trans class Shell_TrayWnd 256'
+            path = $startupPath
+        })
+    }
+# error handling
+    else {
+        if (!$checked.Nircmd) {
+            Out-Error -msg 'Missing Nircmd' -escnum 1
+        }
+    }
+
+# powershell + terminal
     [void]$junctions.Add(@{
         path = "$env:LOCALAPPDATA\Packages\Microsoft.WindowsTerminal*\LocalState"
         target = "$env:USERPROFILE\.config\terminal"
     })
-
-# add PowerShell user folder to $junctions
     [void]$junctions.Add(@{
-        path = "$env:USERPROFILE\Documents\PowerShell"
-        target = "$env:USERPROFILE\Documents\WindowsPowerShell"
+        path = "$env:USERPROFILE\Documents\WindowsPowerShell"
+        target = "$env:USERPROFILE\Documents\PowerShell"
     })
+    Copy-Item -Path $PSScriptRoot\powershell\* -Destination $env:USERPROFILE\Documents\PowerShell -Recurse -Force
+# oh-my-posh
+    if ($config.prompt -eq 'omp' -and $checked.Posh) {
+        $null = New-Item -Path $env:USERPROFILE\Documents\PowerShell\config.json -Value '{"prompt": 1}' -Force
+    }
+# starship
+    elseif ($config.prompt -eq 'starship' -and $checked.Starship) {
+        $null = New-Item -Path $env:USERPROFILE\Documents\PowerShell\config.json -Value '{"prompt": 2}' -Force
+    }
+
+# packages
+<#
+    $config.packages.foreach($item ,{
+        switch ($item) {
+            glazewm {
+            	
+            }
+            nvim {
+
+            }
+            scoop {
+
+            }
+            nvim {
+
+            }
+            winfetch {
+
+            }
+        }
+    })
+#>
 
 # create junctions for every item in $junctions
-    foreach ($item in $junctions) {
-        & $createJunction @item
-    }
+foreach ($item in $junctions) {
+    New-Junction @item
+}
 
 # create shortcut for every item in $shortcuts
-    foreach ($item in $shortcuts) {
-        if (!(Test-Path -Path "$startupPath\$($item.name).lnk")) {
-            & $createLink @item
-        } else {
-            $null = Remove-Item -Path "$startupPath\$($item.name).lnk"
-            & $createLink @item
-        }
-    }
+foreach ($item in $shortcuts) {
+    New-Shortcut @item
+}
+
+<#
+# re route the nvim config folder to .config\nvim
+    [void]$junctions.Add(@{
+        path = "$env:LOCALAPPDATA\nvim"
+        target = "$env:USERPROFILE\.config\nvim"
+    })
 
 # copy items to their destination
-    Copy-Item -Path $PSScriptRoot\powershell\* -Destination $env:USERPROFILE\Documents\WindowsPowerShell\ -Recurse -Force
+    Copy-Item -Path $PSScriptRoot\powershell\* -Destination $env:USERPROFILE\Documents\PowerShell\ -Recurse -Force
     Copy-Item -Path $PSScriptRoot\.config\* -Destination $env:USERPROFILE\.config -Recurse -Force 
+#>
